pyproject.toml
[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::FutureWarning:"
]
addopts = "--doctest-modules --ignore=libs --ignore=docs"

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310']
string_normalization = true
magic_trailing_comma = true

[tool.isort]
profile = "black"
line_length = 120
include_trailing_comma = true
extend_skip = ["__init__.py"]

[tool.flake8]
ignore = ['E231', 'E241', 'W503', 'E203','E501', 'E741', 'F841']
per-file-ignores = [
    '__init__.py:F401, F403',
]
max-line-length = 120
count = true
exclude = ['docs/**', '.tox/**', '.venv/**']

[tool.poetry]
name = "zgoubidoo"
version = "2023.1"
description = "Zgoubidoo: a modern Python 3 interface to the Zgoubi particle tracking code"
authors = ["Cédric Hernalsteens <cedric.hernalsteens@ulb.be>",
           "Robin Tesse <robin.tesse@ulb.be>",
           "Marion Vanwelde <marion.vanwelde@ulb.be>"]

[tool.poetry.dependencies]
python = ">3.8.1 <3.11"
parse = "^1.19.0"
lmfit = "^1.0.2"
mypy = "^0.991"
nbstripout = "^0.4.0"
numpy-stl = "^2.16.0"
sympy = "^1.10.1"

georges-core = { path = "libs/georges-core/", develop = true }

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = {extras = ["jupyter"], version = "^22.10.0"}
pylint = "^2.15.2"
flake8 = "^6.0.0"
isort = "^5.10.1"
flake8-pyproject = "^1.2.2"
mypy = "^0.991"
coverage = {extras = ["toml"], version = "^6.5.0"}
pre-commit = "^2.21.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^5.3.0"
sphinx-material =  "^0.0.35"
sphinx-automodapi =  "^0.14.1"
sphinxcontrib-bibtex =  "^2.2.0"
sphinx-copybutton =  "^0.5.1"
sphinx-autobuild =  "^2021.3.14"
sphinx-rtd-theme = "^1.1.1"
m2r2 =  "^0.2.8"
pandoc =  "^1.1.0"
nbsphinx =  "^0.8.2"
jupyter-sphinx = "^0.4.0"
jupyter-server-mathjax = "^0.2.6"
cairosvg = "^2.5.2"
sphinxcontrib-svg2pdfconverter = {extras = ["cairosvg"], version = "^1.2.1"}
sphinx-tabs = "^3.4.0"

[tool.coverage.run]
branch = true
relative_files = true
source = ["georges_core"]

[tool.coverage.report]
skip_covered = true
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    'raise NotImplementedError',
    'except KeyError',
    'if 0:',
    'if __name__ == .__main__.:',
    'class .*\bProtocol\):',
    '@(abc\.)?abstractmethod',
    ]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
